feat(orders): funções de atualização de pedidos e items

+++ APIRestaurante.Application/DTOs/UpdateMenuItemDTO.cs
@@ -3,5 +3,5 @@ namespace APIRestaurante.Application.DTOs;
 public class MenuItemUpdateDto
 {
     public string? Name { get; set; }
-    public decimal? PriceCents { get; set; }
+    public int? PriceCents { get; set; }
 }
